Fix Vite HMR breakage in Docker container
The value of server.hmr.host defaults to localhost.  This breaks HMR
when Vite dev server is running concurrently with PHP development server
in a Docker container hosted on a remote Linux server.

This is a bug and it is 100% reproducible in the above environment.  Yet
it is very easy to fix by setting server.hmr.host to a valid URI rather
than localhost.